% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigexp.R
\name{print.bigexp_spec}
\alias{print.bigexp_spec}
\title{Print method for bigexp_spec objects}
\usage{
\method{print}{bigexp_spec}(x, ...)
}
\arguments{
\item{x}{A "bigexp_spec" object.}

\item{...}{Unused.}
}
\description{
This print method shows a compact summary of the expansion settings and the
predictors that are treated as continuous or categorical.
}
