% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.svem_significance_test}
\alias{print.svem_significance_test}
\title{Print Method for SVEM Significance Test}
\usage{
\method{print}{svem_significance_test}(x, ...)
}
\arguments{
\item{x}{An object of class \code{svem_significance_test}.}

\item{...}{Additional arguments (not used).}
}
\description{
Prints the p-value from an object of class \code{svem_significance_test}.
}
