% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svem_thompson_batch.R
\name{print.svem_thompson_batch}
\alias{print.svem_thompson_batch}
\title{Print method for svem_thompson_batch objects}
\usage{
\method{print}{svem_thompson_batch}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"svem_thompson_batch"}.}

\item{...}{Unused.}
}
\value{
\code{x}, invisibly.
}
\description{
Print method for svem_thompson_batch objects
}
