% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigexp.R
\name{with_bigexp_contrasts}
\alias{with_bigexp_contrasts}
\title{Evaluate an expression with the spec's recorded contrast options}
\usage{
with_bigexp_contrasts(spec, code)
}
\arguments{
\item{spec}{A "bigexp_spec".}

\item{code}{Code to evaluate with temporarily restored \code{options(contrasts=...)}.}
}
\description{
Evaluate an expression with the spec's recorded contrast options
}
\examples{
# with_bigexp_contrasts(spec, { fit <- SVEMnet(f, data = d, ...) })
}
