% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigexp.R
\name{bigexp_formula}
\alias{bigexp_formula}
\title{Construct a formula for a new response using a bigexp_spec}
\usage{
bigexp_formula(spec, response)
}
\arguments{
\item{spec}{A "bigexp_spec".}

\item{response}{Character scalar: the new response (column name in your data).}
}
\value{
A formula like \verb{response ~ <locked big expansion>}.
}
\description{
Useful when you want to fit multiple responses with the same expansion.
}
\examples{
# f2 <- bigexp_formula(spec, "y2")
}
