% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigexp.R
\name{bigexp_model_matrix}
\alias{bigexp_model_matrix}
\title{Build a model matrix using the spec's stored contrasts (if present)}
\usage{
bigexp_model_matrix(spec, data)
}
\arguments{
\item{spec}{A "bigexp_spec".}

\item{data}{A data frame to prepare and encode.}
}
\value{
The design matrix returned by \code{model.matrix()}.
}
\description{
Build a model matrix using the spec's stored contrasts (if present)
}
\examples{
# MM <- bigexp_model_matrix(spec, newdata)
}
