% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.R
\name{coef.svem_model}
\alias{coef.svem_model}
\title{Plot Coefficient Nonzero Percentages from a SVEMnet Model}
\usage{
\method{coef}{svem_model}(object, ...)
}
\arguments{
\item{object}{An object of class \code{svem_model} returned by the \code{SVEMnet} function.}

\item{...}{other arguments to pass.}
}
\value{
Invisibly returns a data frame containing the percentage of bootstraps
        where each coefficient is nonzero.
}
\description{
This function calculates the percentage of bootstrap iterations in which each coefficient
is nonzero.
}
\section{Acknowledgments}{

Development of this package was assisted by GPT o1-preview, which helped in constructing the structure of some of the code and the roxygen documentation. The code for the significance test is taken from the supplementary material of Karl (2024) (it was handwritten by that author).
}

