% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.svemnet.R
\name{plot.svem_model}
\alias{plot.svem_model}
\title{Plot Method for SVEM Models}
\usage{
\method{plot}{svem_model}(x, plot_debiased = FALSE, ...)
}
\arguments{
\item{x}{An object of class \code{svem_model}.}

\item{plot_debiased}{Logical; if \code{TRUE}, includes debiased predictions if available.}

\item{...}{Additional aesthetics passed to \code{geom_point()}.}
}
\value{
A \code{ggplot} object.
}
\description{
Plots actual versus predicted values for an \code{svem_model}.
}
